+++ src/redturtle/prenotazioni/adapters/conflict.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
 from datetime import timedelta
+
+import six
 from plone.memoize.instance import memoize
 from Products.CMFCore.utils import getToolByName
-from redturtle.prenotazioni.adapters.slot import BaseSlot
 from six.moves import range
 from zope.component import Interface
 from zope.interface import implementer
 
-import six
+from redturtle.prenotazioni.adapters.slot import BaseSlot
 
 
 class IConflictManager(Interface):