+++ src/redturtle/prenotazioni/content/validators.py
@@ -1,16 +1,19 @@
 # -*- coding: utf-8 -*-
+import calendar
+import json
 from datetime import date
-from redturtle.prenotazioni import _
-from redturtle.prenotazioni.adapters.slot import interval_is_contained
-from redturtle.prenotazioni.adapters.slot import is_intervals_overlapping
+
 from z3c.form import validator
 from zope.globalrequest import getRequest
 from zope.i18n import translate
 from zope.interface import Invalid
 from zope.schema import ValidationError
 
-import calendar
-import json
+from redturtle.prenotazioni import _
+from redturtle.prenotazioni.adapters.slot import (
+    interval_is_contained,
+    is_intervals_overlapping,
+)
 
 
 def get_dgf_values_from_request(request, fieldname, columns=[]):