+++ src/redturtle/prenotazioni/exceptions/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .booker import BookerException, BookingsLimitExceded  # noqa