+++ src/redturtle/prenotazioni/scripts/io_tools/__init__.py
@@ -5,10 +5,9 @@ __author__ = """Mauro Amico"""
 __email__ = "mauro.amico@gmail.com"
 __version__ = "0.1.0"
 
-from . import monkey
-
 import logging
 
+from . import monkey
 
 logger = logging.getLogger(__name__)
 monkey.apply()