+++ src/redturtle/prenotazioni/scripts/io_tools/rdbms.py
@@ -1,8 +1,8 @@
 # -*- encoding: utf-8 -*-
 from datetime import datetime
+
 from io_tools.storage import Storage as Base
 from pony import orm
-
 
 db = orm.Database()