+++ src/redturtle/prenotazioni/tests/test_booking_notify_on_confirm_by_endoint.py
@@ -1,22 +1,19 @@
 # -*- coding: UTF-8 -*-
+import email
+import unittest
+from datetime import date, datetime, timedelta
 
-from datetime import date
-from datetime import datetime
-from datetime import timedelta
+import pytz
 from plone import api
-from plone.app.testing import setRoles
-from plone.app.testing import TEST_USER_ID
+from plone.app.testing import TEST_USER_ID, setRoles
+from zope.globalrequest import getRequest
+from zope.interface import implementer
+from zope.interface.interfaces import IObjectEvent
+
 from redturtle.prenotazioni.adapters.booker import IBooker
 from redturtle.prenotazioni.testing import (
     REDTURTLE_PRENOTAZIONI_API_INTEGRATION_TESTING,
 )
-from zope.interface import implementer
-from zope.interface.interfaces import IObjectEvent
-from zope.globalrequest import getRequest
-
-import email
-import pytz
-import unittest
 
 
 @implementer(IObjectEvent)