+++ src/redturtle/prenotazioni/tests/test_prenotazione_events.py
@@ -1,20 +1,18 @@
 # -*- coding: UTF-8 -*-
-from datetime import date
-from datetime import datetime
-from datetime import timedelta
+import email
+import unittest
+from datetime import date, datetime, timedelta
+
+import pytz
 from plone import api
-from plone.app.testing import setRoles
-from plone.app.testing import TEST_USER_ID
-from redturtle.prenotazioni.adapters.booker import IBooker
-from redturtle.prenotazioni.prenotazione_event import MovedPrenotazione
-from redturtle.prenotazioni.testing import REDTURTLE_PRENOTAZIONI_FUNCTIONAL_TESTING
+from plone.app.testing import TEST_USER_ID, setRoles
 from zope.event import notify
 from zope.interface import implementer
 from zope.interface.interfaces import IObjectEvent
 
-import email
-import pytz
-import unittest
+from redturtle.prenotazioni.adapters.booker import IBooker
+from redturtle.prenotazioni.prenotazione_event import MovedPrenotazione
+from redturtle.prenotazioni.testing import REDTURTLE_PRENOTAZIONI_FUNCTIONAL_TESTING
 
 
 @implementer(IObjectEvent)