+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 2.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Remove Contributor from the package permissions map
+  [folix-01]
 
 
 2.0.0 (2023-09-12)