+++ CHANGES.rst
@@ -5,6 +5,13 @@ Changelog
 2.0.1 (unreleased)
 ------------------
 
+- Rimosso searchabletext di prenotazioni doppio
+  [mamico]
+
+- Aggiunto indexer per fiscalcode uppercase per
+  fare ricerche case insensitive
+  [mamico]
+
 - Remove Contributor from the package permissions map
   [folix-01]