+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changelog
   [mamico]
 
 - Remove Contributor from the package permissions map
+- Add configurable simultaneous bookings limit for the same user.
   [folix-01]